requirements.txt
asgiref==3.8.1
attrs==24.2.0
Basket==1.0
black==24.4.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
Django==4.2.14
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
flake8==7.1.0
idna==3.7
inflection==0.5.1
isort==5.13.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
pillow==10.4.0
pip-system-certs==4.0
platformdirs==4.2.2
pycodestyle==2.12.0
pyflakes==3.2.0
PyJWT==2.8.0
python-dotenv==1.0.1
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
sqlparse==0.5.0
stripe==10.7.0
tomli==2.0.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
wrapt==1.16.0
